update routing
Main page is now `/` and login button redirects to `/login`. This allows to provide further login information and options.
