+++ web/src/app/components/login-button/login-button.component.ts
@@ -14,6 +14,5 @@ export class LoginButtonComponent implements OnInit {
 
   onLogin() {
     this.login.emit();
-    window.location.assign('/api/auth/login');
   }
 }
