--- web/src/app/routes/login/login.component.scss
@@ -1,235 +1,0 @@
-@import '../../../sass/vars';
-@import '../../../sass/mixins';
-
-$spacer-color: rgba(white, 0.25);
-
-@keyframes btn-features-inner {
-  0% {
-    opacity: 0.75;
-    clip-path: polygon(
-      0% -80%,
-      50% -40%,
-      100% -80%,
-      100% -40%,
-      50% 0%,
-      0% -40%
-    );
-  }
-  100% {
-    opacity: 0.75;
-    clip-path: polygon(
-      0% 100%,
-      50% 140%,
-      100% 100%,
-      100% 140%,
-      50% 180%,
-      0% 140%
-    );
-  }
-}
-
-@keyframes btn-start-inner {
-  0% {
-    opacity: 0.75;
-    clip-path: polygon(
-      -20% 0%,
-      -10% 0%,
-      0% 50%,
-      -10% 100%,
-      -20% 100%,
-      -10% 50%
-    );
-  }
-  100% {
-    opacity: 0.75;
-    clip-path: polygon(
-      100% 0,
-      110% 0,
-      120% 50%,
-      110% 100%,
-      100% 100%,
-      110% 50%
-    );
-  }
-}
-
-hr {
-  width: 90vw;
-  max-width: 1300px;
-  border-color: $spacer-color !important;
-}
-
-.login-container {
-  position: absolute;
-  top: 0px;
-  left: 0px;
-  right: 0px;
-}
-
-.header-box {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  height: calc(100vh + 5vh);
-  background-image: url('/assets/landingpage/background.png');
-  background-position: 50%;
-  background-repeat: no-repeat;
-  background-size: cover;
-  clip-path: polygon(0 0, 100% 0, 100% 95%, 0 100%);
-
-  > * {
-    margin: 0 auto;
-  }
-
-  > .logo {
-    margin: 30vh auto 10vh auto;
-
-    .header-image {
-      width: 600px;
-      max-width: 80vw;
-    }
-  }
-
-  > .button {
-    display: flex;
-
-    > button {
-      position: relative;
-      background-color: transparent;
-      border: solid 1px white;
-      border-radius: 0px !important;
-      cursor: pointer;
-
-      &:first-child {
-        margin-right: 20px;
-      }
-
-      &::before {
-        content: '';
-        position: absolute;
-        top: 0;
-        left: 0;
-        width: 100%;
-        height: 100%;
-        z-index: -1;
-        background-color: #ffffff;
-        opacity: 0;
-        transition: all 0.25s ease;
-      }
-
-      &:first-child:hover::before {
-        animation: btn-features-inner 1s ease;
-      }
-
-      &:last-child:hover::before {
-        animation: btn-start-inner 1s ease;
-      }
-    }
-  }
-}
-
-.exmaple-tile {
-  display: grid;
-  width: 100%;
-  grid-template-columns: 50% 50%;
-  box-sizing: border-box;
-  align-items: center;
-
-  > * {
-    padding: 0 20px;
-  }
-
-  > img {
-    width: 100%;
-    max-width: 500px;
-  }
-
-  &:nth-child(odd) > img {
-    margin-left: auto;
-  }
-
-  > div {
-    max-width: 480px;
-    h3 {
-      margin-bottom: 20px !important;
-    }
-  }
-
-  &:nth-child(even) > div {
-    margin-left: auto;
-  }
-}
-
-.more-features {
-  text-align: center;
-  margin: 100px 0px;
-
-  > * {
-    padding: 10px 0px;
-  }
-}
-
-.getting-started-grid {
-  display: grid;
-  grid-template-columns: 50% 3px 50%;
-  width: 100%;
-  max-width: 1000px;
-  margin: 80px auto;
-
-  @media only screen and (max-width: 1000px) {
-    grid-template-columns: 100%;
-    .self-host {
-      margin-top: 50px;
-    }
-  }
-
-  > div.side {
-    padding: 0px 80px;
-    > h3 {
-      text-align: center;
-      padding-bottom: 40px;
-    }
-  }
-
-  > .mid-spacer {
-    width: 100%;
-    height: 100%;
-    background-color: $spacer-color;
-  }
-
-  .inv-button {
-    width: 100%;
-    display: block;
-    margin: 0 auto 20px auto;
-    box-sizing: border-box;
-  }
-
-  .text-embed {
-    font-family: Courier New, Courier, monospace;
-    background-color: rgba(black, 0.25);
-    padding: 18px;
-  }
-}
-
-.link-list {
-  display: flex;
-  justify-content: center;
-
-  > a {
-    display: flex;
-    align-items: center;
-    border: solid 1px white;
-    padding: 10px;
-    margin: 40px 8px;
-    @include hover-btn;
-    > img {
-      margin-right: 10px;
-    }
-  }
-}
-
-.footer {
-  margin-top: 50px;
-  text-align: center;
-  font-size: 14px;
-}
