+++ web/src/app/routes/start/start.component.html
@@ -1,0 +1,132 @@
+<div class="login-container">
+  <div class="header-box">
+    <div class="logo">
+      <a>
+        <img class="header-image" src="/assets/landingpage/banner.png" />
+      </a>
+    </div>
+    <div class="button">
+      <button (click)="scrollTo('start-features')">DISCOVER FEATURES</button>
+      <button (click)="scrollTo('start-getting-started')">
+        GETTING STARTED
+      </button>
+    </div>
+  </div>
+
+  <a id="start-features"></a>
+
+  <!-- REPORT SYSTEM -->
+  <div class="exmaple-tile">
+    <img src="/assets/landingpage/500px/ex-report.png" />
+    <div>
+      <h3>REPORT SYSTEM</h3>
+      <p>
+        A simple to use system to keep track of user reports, bans, kicks and
+        chat-mutes.
+      </p>
+    </div>
+  </div>
+  <!-- GUILD BACKUPS -->
+  <div class="exmaple-tile">
+    <div>
+      <h3>GUILD BACKUPS</h3>
+      <p>
+        A fully automatic system which records the current guild state every 12
+        hours. Just pick a backup and recover it or download it as JSON file.
+      </p>
+    </div>
+    <img src="/assets/landingpage/500px/ex-backups.png" />
+  </div>
+  <!-- KARMA SYSTEM -->
+  <div class="exmaple-tile">
+    <img src="/assets/landingpage/500px/ex-karma.png" />
+    <div>
+      <h3>KARMA SYSTEM</h3>
+      <p>
+        Honor users which post high quality content by giving them karma points
+        via reactions like 👍 or punishing low quality content by lowering their
+        karma using a reaction like 👎.
+        <br /><br />
+        Of course, caps and emotes are fully customizable.
+      </p>
+    </div>
+  </div>
+  <!-- CODE EXEC-->
+  <div class="exmaple-tile">
+    <div>
+      <h3>CODE EXECUTION</h3>
+      <p>
+        shinpuru detects code embeds in chat and by clicking the ▶ reaction, you
+        can compile and run the code. The result is then posted into the chat
+        afterwards.
+        <br /><br />
+        The code is savely executed by the
+        <a class="link" href="https://www.jdoodle.com" target="_blank"
+          >JDoodle</a
+        >
+        API.
+      </p>
+    </div>
+    <img src="/assets/landingpage/500px/ex-codeexec.png" />
+  </div>
+  <div class="more-features">
+    <h2>AND A BUNCH OF MORE FEATURES</h2>
+    <h3>READY TO BE DISCOVERED</h3>
+  </div>
+
+  <hr />
+
+  <a id="start-getting-started"></a>
+
+  <div class="getting-started-grid">
+    <div class="side">
+      <h3>INVITE</h3>
+      <a *ngIf="info?.localinvite" [href]="info?.localinvite">
+        <button class="inv-button">THIS INSTANCE</button>
+      </a>
+      <a *ngIf="info?.publicmaininvite" [href]="info?.publicmaininvite">
+        <button class="inv-button bg-green">PUBLIC MAIN INSTANCE</button>
+      </a>
+      <a *ngIf="info?.publiccaranyinvite" [href]="info?.publiccaranyinvite">
+        <button class="inv-button bg-orange">PUBLIC CANARY INSTANCE</button>
+      </a>
+    </div>
+    <div class="mid-spacer"></div>
+    <div class="side self-host">
+      <h3>SELF HOST</h3>
+      <p class="text-embed">$ docker pull zekro/shinpuru</p>
+      <a href="https://github.com/zekroTJA/shinpuru/wiki/docker">
+        <button class="inv-button">WITH DOCKER</button>
+      </a>
+    </div>
+  </div>
+
+  <hr />
+
+  <a id="start-links"></a>
+
+  <div class="link-list">
+    <a href="https://github.com/zekroTJA/shinpuru" target="_blank">
+      <img width="25" src="/assets/landingpage/github.svg" />
+      <span>CONTRIBUTE</span>
+    </a>
+    <a href="https://ko-fi.com/zekro" target="_blank">
+      <img width="25" src="/assets/landingpage/ko-fi.svg" />
+      <span>SUPPORT</span>
+    </a>
+  </div>
+
+  <hr />
+
+  <div class="footer">
+    <p>
+      © {{ currentYear }} Ringo Hoffmann (zekro Development).<br />
+      Covered by the
+      <a
+        class="link"
+        href="https://github.com/zekroTJA/shinpuru/blob/master/LICENCE"
+        >MIT License</a
+      >.
+    </p>
+  </div>
+</div>
